backend/communications/api/v1/serializers.py
from rest_framework import serializers
from communications.models import Notification


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = "__all__"
